Dals splitted in two entities: roomRepository and sessionRepository
