+++ back/src/dals/room/index.ts
@@ -1,2 +1,3 @@
 export * from './room.model';
 export * from './room.context';
+export * from './repository';
