+++ back/src/dals/room/repository/room.contract.ts
@@ -1,0 +1,8 @@
+import { RoomInfo } from 'dals';
+
+export interface RoomRepositoryContract {
+  saveRoomInfo: (newRoomInfo: RoomInfo, action: string)  => Promise<void>;
+  isRoomAvailable: (room: string) => Promise<boolean>;
+  getRoomContent:(room: string) => Promise<string>;
+}
+
