+++ back/src/dals/session/repository/session.contract.ts
@@ -1,14 +1,8 @@
-import { RoomInfo } from 'dals/room';
-import { ConnectSessionInfo } from 'dals/session';
-import { UserSession } from '../session.model';
+import { UserSession } from 'dals';
 
-export interface SessionRepository {
+export interface SessionRepositoryContract {
   isTrainerUser: (connectionId: string) => Promise<boolean>;
-  saveRoomInfo: (newRoomInfo: RoomInfo, action: string)  => Promise<void>;
   isExistingConnection: (connectionId: string) => Promise<boolean>;
-  isRoomAvailable: (room: string) => Promise<boolean>;
   addNewUser: (userSession: UserSession) => Promise<void>;
   getRoomFromConnectionId: (connectionId: string) => Promise<string>;
-  getRoomContent:(room: string) => Promise<string>;
 }
-
